Cast B_OK to (uint32) to solve an overloading ambiguity (I don't agree with the compiler, though :-)), I'm about to introduce.
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@25384 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/apps/sudoku/SudokuView.cpp
@@ -411,7 +411,7 @@ SudokuView::SaveTo(BDataIO &stream, uint32 as)
 				delete view;
 				bitmap->Unlock();
 			}
-			BMessage msg(B_OK);
+			BMessage msg((uint32)B_OK);
 			status = bitmap->Archive(&msg);
 			if (status >= B_OK) {
 				status = msg.Flatten(&stream);
